Clarify using Default Key Policy
There is a mismatch in documentation where its explained that using the Default Key Policy requires calling the KMS CreateKey API without any explicit key policy. However, because of the way the `KeyPolicy` field works in CloudFormation, if your intention is to use the Default Key Policy, you can't actually pass an empty `KeyPolicy` field. The solution to this is to copy the default key policy that would normally be created for you during a regular CreateKey API Action into the CloudFormation `KeyPolicy` field.